fix: fix hapi-swagger crash due to response validation not done by Joi
Hapi-swagger is generating the swagger.json based on the routes informations and Joi validations.
Since the HapiJS 20 upgrade, HapiJS does not use anymore Joi for internal validation but an internal light version of Joi.
The thing is when hapi-swagger is generating the documentation, if different validation libs or versions are used it breaks.
I figured out that we had 2 places where we did not use the Joi validation that needed to be changed.