feat(quotes api) unify quotes api endpoint returned ressource
There was a difference between the ressource returned by the POST /quotes and the GET/PUT /quotes so we unified the returned ressource.
A field is acutally missing and will be added in next commits is the special_operation_code and special_operation_code_applied_at.
We also added some alloNull on ressource fileds that could be null.