fix(quotes api) do not return risk.person or risk.otherPeople if firstname and lastname are null
Actually when we save a quote, if there is no risk.person and risk.otherPeople we still create empty entities in the database.
I think this has been done to facilitate the update (not sure but it's my guess).
But the empty entites are returned via the API, so we quick fixed the code so that it does not return the entities in this case.
A more complete fix should be to work on the repository save/update so that entities are not created if empty.